% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructor.R
\name{add_annotation}
\alias{add_annotation}
\title{Add annotations to the DAG object}
\usage{
add_annotation(dag, annotation)
}
\arguments{
\item{dag}{An \code{ontology_DAG} object.}

\item{annotation}{A list of character vectors which contain items annotated to the terms. Names of the list should be the term names. In the DAG, items
annotated to a term will also be annotated to its parents. Such merging
is applied automatically in the package.}
}
\value{
An \code{ontology_DAG} object.
}
\description{
Add annotations to the DAG object
}
