% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_global.R
\docType{data}
\name{ALL_GROUP_SIM_METHODS}
\alias{ALL_GROUP_SIM_METHODS}
\title{All supported group similarity methods}
\format{
An object of class \code{character} of length 28.
}
\usage{
ALL_GROUP_SIM_METHODS
}
\value{
A vector all supported grouped similarity methods.
}
\description{
All supported group similarity methods
}
\keyword{datasets}
