% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_global.R
\docType{data}
\name{ALL_TERM_SIM_METHODS}
\alias{ALL_TERM_SIM_METHODS}
\title{All supported similarity methods}
\format{
An object of class \code{character} of length 33.
}
\usage{
ALL_TERM_SIM_METHODS
}
\value{
A vector all supported similarity methods.
}
\description{
All supported similarity methods
}
\keyword{datasets}
