% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructor.R
\name{create_ontology_DAG_from_igraph}
\alias{create_ontology_DAG_from_igraph}
\title{Create the ontology_DAG object from the igraph object}
\usage{
create_ontology_DAG_from_igraph(
  g,
  relations = NULL,
  verbose = simona_opt$verbose
)
}
\arguments{
\item{g}{An \code{\link[igraph:aaa-igraph-package]{igraph::igraph}} object.}

\item{relations}{A vector of relation types. The length of the vector should be the same as the number of edges in \code{g}.}

\item{verbose}{Whether to print messages.}
}
\value{
An \code{ontology_DAG} object.
}
\description{
Create the ontology_DAG object from the igraph object
}
