% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructor.R
\name{createOntologyDAG}
\alias{createOntologyDAG}
\title{Create the OntologyDAG object}
\usage{
createOntologyDAG(
  parents,
  children,
  relations = NULL,
  source = "Ontology",
  annotation = NULL
)
}
\arguments{
\item{parents}{A vector of parent terms.}

\item{children}{A vector of child terms.}

\item{relations}{A vector of parent-child relations.}

\item{source}{Source of the ontology.}

\item{annotation}{A list of annotations. The names in the list should correspond to the term names.}
}
\description{
Create the OntologyDAG object
}
