% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term.R
\name{dag_depth}
\alias{dag_depth}
\alias{dag_height}
\title{Maximal depth and height}
\usage{
dag_depth(dag, use_cache = TRUE)

dag_height(dag, use_cache = TRUE)
}
\arguments{
\item{use_cache}{Only for testing purpose.}
}
\description{
Maximal depth and height
}
\details{
aa
}
