% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich.R
\name{dag_enrich_terms_by_permutation}
\alias{dag_enrich_terms_by_permutation}
\title{Enrichment analysis on offspring terms by permutation test}
\usage{
dag_enrich_terms_by_permutation(dag, value, n = 1000)
}
\arguments{
\item{dag}{An \code{ontology_DAG} object.}

\item{value}{A numeric value. The value should correspond to terms in \code{dag@terms}.}

\item{n}{Number of permutations.}
}
\value{
A data frame with the following columns:
\itemize{
\item \code{term}: Term names.
\item \code{stats}: The statistics of terms.
\item \code{n_offspring}: Number of offspring terms of \code{t} (including \code{t} itself).
\item \code{z_score}: Defined as \code{(s - mean)/sd} where \code{mean} and \code{sd} are calculated from random permutation.
\item \code{log2_fold_enrichment}: defined as \code{log2(s/mean)} where \code{mean} is calculated from random permutation.
\item \code{p_value}: P-values from permutation test.
\item \code{p_adjust}: Adjusted p-values from the BH method.
}

The number of rows in the data frame is the same as the number of terms in the DAG.
}
\description{
Enrichment analysis on offspring terms by permutation test
}
\details{
The function tests whether a term \code{t}'s offspring terms have an over-represented pattern on values in \code{value}.
Denote \code{T} as the set of \code{t}'s offspring terms plus \code{t} itself, and \code{v} as the numeric vector of \code{value}, we first
calculate a score \code{s} based on values in \code{T}:

\if{html}{\out{<div class="sourceCode">}}\preformatted{s = mean_\{terms in T\}(v)
}\if{html}{\out{</div>}}

To construct a random version of \code{s}, we randomly sample \code{n_T} terms from the DAG where \code{n_T} is the size of set \code{T}:

\if{html}{\out{<div class="sourceCode">}}\preformatted{sr_i = mean_\{n_T randomly sampled terms\}(v)
}\if{html}{\out{</div>}}

where index \code{i} represents the i^th sampling. If we sample \code{k} times, the p-value is calculated as:

\if{html}{\out{<div class="sourceCode">}}\preformatted{p = sum_\{go over k\}(I(sr > s))/k
}\if{html}{\out{</div>}}
}
\examples{
\dontrun{
dag = create_ontology_DAG_from_GO_db() 
value = runif(dag_n_terms(dag))
df = dag_enrich_terms_by_permutation(dag, value)
}
1
}
