% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructor.R
\name{dag_has_terms}
\alias{dag_has_terms}
\title{Whether the terms exist in the DAG}
\usage{
dag_has_terms(dag, terms)
}
\arguments{
\item{dag}{An \code{ontology_DAG} object.}

\item{terms}{A vector of term IDs.}
}
\value{
A logical vector.
}
\description{
Whether the terms exist in the DAG
}
