% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transverse.R
\name{dag_longest_dist_to_offspring}
\alias{dag_longest_dist_to_offspring}
\alias{dag_shortest_dist_to_offspring}
\alias{dag_longest_dist_from_ancestors}
\alias{dag_shortest_dist_from_ancestors}
\title{Distance from terms in the DAG}
\usage{
dag_longest_dist_to_offspring(dag, from, terms = NULL, background = NULL)

dag_shortest_dist_to_offspring(dag, from, terms = NULL, background = NULL)

dag_longest_dist_from_ancestors(dag, to, terms = NULL, background = NULL)

dag_shortest_dist_from_ancestors(dag, to, terms = NULL, background = NULL)
}
\arguments{
\item{dag}{An \code{ontology_DAG} object.}

\item{from}{A single term or a vector of terms. For all downstream terms, the maximal or the minimal distance to the terms specified in this variable will be taken.}

\item{terms}{A vector of term names. If it is set, the returned vector will be subsetted to the terms that have been set here.}

\item{background}{A vector of terms. Then the lookup will only be applied in this set of terms.}

\item{to}{Same format as the \code{from} argument.}
}
\description{
Distance from terms in the DAG
}
\details{
As a special case, when \code{from} is the root term, \code{dag_longest_dist_to_offspring()} is the same as \code{dag_depth()},
and when \code{to} are all leaf terms, \code{dag_shortest_dist_to_offspring()} is the same as \code{dag_height()}.
}
