% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{dag_random}
\alias{dag_random}
\title{Generate a random DAG}
\usage{
dag_random(n_children = c(2, 5), max = 1000, p = 0.1, power = 0.5)
}
\arguments{
\item{n_children}{Range of the numbers of children on a term to sample.}

\item{max}{Maximal number of terms.}

\item{p}{Probability to connect to other terms.}

\item{power}{The power.}
}
\value{
An \code{ontology_DAG} object.
}
\description{
Generate a random DAG
}
\details{
It first generates a tree growing from the root. On each node, it has
a random number of children sampled from \code{n_children} and each child node
can extend with a probability of \code{1/(height^power)}.

In the second step, each node on the tree has a probability of \code{p} to link to other
terms in the tree.

The growing stops when the number of terms reaches \code{max} or there is no term can further extend.
}
\examples{
dag_random()
dag_random(p = 0)  # a tree
}
