% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructor.R
\name{dag_root}
\alias{dag_root}
\alias{dag_leaves}
\title{Root or leaves of the DAG}
\usage{
dag_root(dag, in_labels = TRUE)

dag_leaves(dag, in_labels = TRUE)
}
\arguments{
\item{dag}{An `ontology_DAG` object.}

\item{in_labels}{Whether the terms are in their names or as the integer indices?}
}
\description{
Root or leaves of the DAG
}
