% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{dag_shiny}
\alias{dag_shiny}
\title{A shiny app for the DAG}
\usage{
dag_shiny(dag)
}
\arguments{
\item{dag}{An \code{ontology_DAG} object.}
}
\description{
A shiny app for the DAG
}
\examples{
if(FALSE) {
    dag = create_ontology_DAG_from_GO_db()
    dag_shiny(dag)
}
}
