% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{group_sim}
\alias{group_sim}
\title{Semantic similarity between two groups of terms}
\usage{
group_sim(dag, group1, group2, method, control = list())
}
\arguments{
\item{dag}{An `ontology_DAG` object.}

\item{group1}{A vector of term names.}

\item{group2}{A vector of term names.}

\item{method}{A group similarity method. All available methods are in [ALL_GROUP_SIM_METHODS].}

\item{control}{A list of parameters passing to individual methods.}
}
\value{
A numeric scalar.
}
\description{
Semantic similarity between two groups of terms
}
