% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructor.R
\name{mcols,ontology_DAG-method}
\alias{mcols,ontology_DAG-method}
\title{Get meta columns on DAG}
\usage{
\S4method{mcols}{ontology_DAG}(x, use.names = TRUE, ...)
}
\arguments{
\item{x}{An \code{ontology_DAG} object.}

\item{use.names}{Please ignore.}

\item{...}{Other argument. For \code{mcols()}, it can be a vector of column names in the meta data frame.}
}
\description{
Get meta columns on DAG
}
