% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructor.R
\name{metadata,ontology_DAG-method}
\alias{metadata,ontology_DAG-method}
\title{Get metadata on DAG}
\usage{
\S4method{metadata}{ontology_DAG}(x, ...)
}
\arguments{
\item{x}{An \code{ontology_DAG} object.}

\item{...}{Other argument. For \code{metadata()}, it can be a vector of column names in the meta data frame.}
}
\description{
Get metadata on DAG
}
