% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructor.R
\name{metadata<-,ontology_DAG-method}
\alias{metadata<-,ontology_DAG-method}
\title{Set metadata on DAG}
\usage{
\S4method{metadata}{ontology_DAG}(x, ...) <- value
}
\arguments{
\item{x}{An \code{ontology_DAG} object.}

\item{...}{Other argument. For \code{metadata()}, it can be a vector of column names in the meta data frame.}

\item{value}{A data frame or a matrix where rows should correspond to terms in \code{x@terms}.}
}
\description{
Set metadata on DAG
}
\examples{
parents  = c("a", "a", "b", "b", "c", "d")
children = c("b", "c", "c", "d", "e", "f")
dag = create_ontology_DAG(parents, children)
metadata(dag) = data.frame(id = letters[1:6], v = 1:6)
metadata(dag)
metadata(dag, "id")
dag
}
