% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{MICA_term}
\alias{MICA_term}
\alias{MICA_IC}
\alias{LCA}
\alias{LCA_term}
\alias{LCA_depth}
\title{MICA and LCA}
\usage{
MICA_term(dag, terms, IC_method)

MICA_IC(dag, terms, IC_method)

LCA(dag, terms, return_termID = FALSE)

LCA_term(dag, termID)

LCA_depth(dag, termID)
}
\description{
MICA and LCA
}
