% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transverse.R
\name{n_offspring}
\alias{n_offspring}
\alias{n_ancestor}
\alias{n_leaves}
\title{Number of terms}
\usage{
n_offspring(dag, term = NULL, use_cache = TRUE)

n_ancestor(dag, term = NULL, use_cache = TRUE)

n_leaves(dag, term = NULL, use_cache = TRUE)
}
\arguments{
\item{use_cache}{}
}
\description{
Number of terms
}
\details{
Leaf nodes have zero `n_leaves` value.
}
