% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructor.R
\docType{class}
\name{ontology_DAG-class}
\alias{ontology_DAG-class}
\alias{ontology_DAG}
\title{The ontology_DAG class}
\description{
This class defines the DAG structure of an ontology.
}
\section{Slots}{

\describe{
\item{\code{lt_parents}}{A list of length \code{n}. Each element is an integer index vector of parent terms.}

\item{\code{lt_children}}{foo}

\item{\code{lt_children_relations}}{foo}

\item{\code{source}}{The source of the ontology. A character scalar only used as a mark of the returned object.}

\item{\code{terms}}{A character vector of length \code{n}, which contains the term names.}

\item{\code{n_terms}}{An integer scalar of the total number of terms in the DAG.}

\item{\code{root}}{An integer index of the root term.}

\item{\code{leaves}}{An integer vector of the indicies of leaf terms.}

\item{\code{tpl_sorted}}{An integer vector of reordered term indices which has been topologically sorted in the DAG.}

\item{\code{tpl_pos}}{The position of the original term in the topologically sorted path (similar as the rank), e.g. the position of term 1 in the sorted path.}

\item{\code{annotation}}{A list of two elements: \code{list} and \code{names}. The \code{list} element contains a list of length \code{n} and each element
there is a vector of integer indices of annotated items. The full list of annotated items is in the \code{names} element.}

\item{\code{term_env}}{An environment which contains various term-level statistics, mainly for cache purpose.}
}}

