% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.ontology_DAG}
\alias{plot.ontology_DAG}
\title{Visualize the DAG}
\usage{
\method{plot}{ontology_DAG}(x, ...)
}
\arguments{
\item{x}{An `ontology_Dag` object.}

\item{...}{Other parguments.}
}
\description{
Visualize the DAG
}
