% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{print.print_source}
\alias{print.print_source}
\title{Print the source}
\usage{
\method{print}{print_source}(x, ...)
}
\arguments{
\item{x}{An object in the \code{print_source} class.}

\item{...}{Other arguments.}
}
\description{
Print the source
}
\details{
Internally used.
}
