% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_ancestor.R
\name{shortest_distances_via_CA}
\alias{shortest_distances_via_CA}
\alias{longest_distances_via_LCA}
\alias{shortest_distances_directed}
\alias{longest_distances_directed}
\title{Distance on the DAG}
\usage{
shortest_distances_via_CA(dag, terms)

longest_distances_via_LCA(dag, terms)

shortest_distances_directed(dag, terms)

longest_distances_directed(dag, terms)
}
\arguments{
\item{dag}{A `ontology_DAG` object.}

\item{terms}{A vector of term names.}
}
\description{
Distance on the DAG
}
\details{
- `shortest_distances_via_CA()`: it is a sum of two distances: d(c, a) + d(c, b)
where c is a common ancestor of a and b and d(c, a) is the shortest distance from c to a.
- `longest_distances_via_LCA()`: here c is the lowest common ancestor (with the largest depth) of a and b,
and d(c, a) is the longest distance from c to a.
- `shortest_distances_directed()`: it calculates d(a, b). The distance is only calculated when a is an ancestor of b.
- `longest_distances_directed()`: it calculates d(a, b). The distance is only calculated when a is an ancestor of b.
}
