% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructor.R
\name{show,ontology_DAG-method}
\alias{show,ontology_DAG-method}
\title{Print the ontology_DAG object}
\usage{
\S4method{show}{ontology_DAG}(object)
}
\arguments{
\item{object}{An \code{ontology_DAG} object.}
}
\description{
Print the ontology_DAG object
}
