% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructor.R
\name{show,OntologyDAG-method}
\alias{show,OntologyDAG-method}
\title{Print the OntologyDAG object}
\usage{
\S4method{show}{OntologyDAG}(object)
}
\arguments{
\item{dag}{A `OntologyDAG` object.}
}
\description{
Print the OntologyDAG object
}
