% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_global.R
\name{simone_opt}
\alias{simone_opt}
\title{Global options}
\usage{
simone_opt(..., RESET = FALSE, READ.ONLY = NULL, LOCAL = FALSE, ADD = FALSE)
}
\arguments{
\item{...}{Ignore.}

\item{RESET}{Ignore.}

\item{READ.ONLY}{Ignore.}

\item{LOCAL}{Ignore.}

\item{ADD}{Ignore.}
}
\description{
Global options
}
\examples{
simone_opt
}
