% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructor.R
\name{[,OntologyDAG,character,missing,ANY-method}
\alias{[,OntologyDAG,character,missing,ANY-method}
\title{Create a sub-DAG}
\usage{
\S4method{[}{OntologyDAG,character,missing,ANY}(x, i, j, ..., drop = FALSE)
}
\arguments{
\item{x}{A `OntologyDAG` object.}

\item{i}{A single term name.}

\item{j}{Ignored.}

\item{...}{Ignored.}

\item{drop}{Ignored.}
}
\description{
Create a sub-DAG
}
