% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructor.R
\name{[[,OntologyDAG,character,missing-method}
\alias{[[,OntologyDAG,character,missing-method}
\title{Create a sub-DAG}
\usage{
\S4method{[[}{OntologyDAG,character,missing}(x, i, j, ...)
}
\arguments{
\item{x}{A `OntologyDAG` object.}

\item{i}{A single term name.}

\item{j}{Ignored.}

\item{...}{Ignored.}
}
\description{
Create a sub-DAG
}
