% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructor.R
\name{[,ontology_DAG,character,missing-method}
\alias{[,ontology_DAG,character,missing-method}
\alias{[[,ontology_DAG,character,missing-method}
\title{Create a sub-DAG}
\usage{
\S4method{[}{ontology_DAG,character,missing}(x, i, j, ..., drop = FALSE)

\S4method{[[}{ontology_DAG,character,missing}(x, i, j, ...)
}
\arguments{
\item{x}{An \code{ontology_DAG} object.}

\item{i}{A single term name.}

\item{j}{Ignored.}

\item{...}{Ignored.}

\item{drop}{Ignored.}
}
\value{
An \code{ontology_DAG} object.
}
\description{
Create a sub-DAG
}
\details{
It returns a sub-DAG taking node \code{i} as the root.
}
