% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{term_annotations}
\alias{term_annotations}
\alias{annotated_terms}
\title{Get term-item associations}
\usage{
term_annotations(dag, terms, return = "list")

annotated_terms(dag, anno, return = "list")
}
\arguments{
\item{dag}{A `ontology_DAG` object.}

\item{terms}{A vector of term names.}

\item{return}{Whether to return as a list or a matrix?}

\item{anno}{A vector of annotated item names.}
}
\value{
A list of annotated items or a list of terms or a binary matrix showing whether the term is annotated with the item.
}
\description{
Get term-item associations
}
