% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{term_sim}
\alias{term_sim}
\title{Semantic similarity}
\usage{
term_sim(dag, terms, method, control = list())
}
\arguments{
\item{dag}{An \code{ontology_DAG} object.}

\item{terms}{A vector of term names.}

\item{method}{A term similarity method. All available methods are in \code{\link{ALL_TERM_SIM_METHODS}}.}

\item{control}{A list of parameters passing to individual methods. See the subsections.}
}
\value{
A numeric symmetric matrix.
}
\description{
Semantic similarity
}
\section{method}{

\subsection{Sim_Ling_1998}{

Paper link: \url{https://dl.acm.org/doi/10.5555/645527.657297}.
}


\subsection{Sim_Resnik_1999}{

Paper link: \url{https://doi.org/10.1613/jair.514}, \url{https://doi.org/10.1186/1471-2105-9-S5-S4}, \url{https://doi.org/10.1186/1471-2105-11-562}, \url{https://doi.org/10.1155/2013/292063}.
}


\subsection{Sim_FaITH_2010}{

\url{https://doi.org/10.1007/978-3-642-17746-0_39}.
}


\subsection{Sim_PS_2008}{

\url{https://doi.org/10.1016/j.datak.2009.06.008}.
}


\subsection{Sim_Relevance_2006}{

\url{https://doi.org/10.1186/1471-2105-7-302}
}


\subsection{Sim_SimIC_2010}{

\url{https://doi.org/10.48550/arXiv.1001.0958}.
}


\subsection{Sim_EISI_2015}{

\url{https://doi.org/10.1016/j.gene.2014.12.062}
}


\subsection{Sim_XGraSM_2013}{

\url{https://doi.org/10.1186/1471-2105-14-284}
}


\subsection{Sim_GraSM_2005}{

\url{https://doi.org/10.1145/1099554.1099658}.
}


\subsection{Sim_AIC_2014}{

\url{https://doi.org/10.1109/tcbb.2013.176}.
}


\subsection{Sim_Zhang_2006}{
}


what is Sim_universal


\subsection{Sim_Wang_2007}{

Paper link: \url{https://doi.org/10.1093/bioinformatics/btm087}.
}


\subsection{Sim_Rada_1989}{

\url{https://doi.org/10.1109/21.24528}
}


\subsection{Sim_Resnik_edge_2005}{

\url{https://doi.org/10.1145/1097047.1097051}
}


\subsection{Sim_Leocock_1998}{

\url{https://ieeexplore.ieee.org/document/6287675}
}


what is Sim_WP_1994


\subsection{Sim_Slimani_2006}{

\url{https://zenodo.org/record/1075130}
}


\subsection{Sim_Shenoy_2012}{

\url{https://doi.org/10.48550/arXiv.1211.4709}.
}


\subsection{Sim_Pekar_2002}{

\url{https://aclanthology.org/C02-1090/}.
}


\subsection{Sim_Stojanovic_2001}{

\url{https://doi.org/10.1145/500737.500762}
}


\subsection{Sim_Wang_edge_2012}{

\url{https://doi.org/10.1186/1477-5956-10-s1-s18}.
}


\subsection{Sim_Zhong_2002}{

\url{https://doi.org/10.1007/3-540-45483-7_8}
}


\subsection{Sim_AlMubaid_2006}{

\url{https://doi.org/10.1109/IEMBS.2006.259235}
}


\subsection{Sim_Li_2003}{

\url{https://doi.org/10.1109/TKDE.2003.1209005}
}


\subsection{Sim_RSS_2013}{

\url{https://doi.org/10.1371/journal.pone.0066745}
}


\subsection{Sim_HRSS_2013}{

\url{https://doi.org/10.1371/journal.pone.0066745}
}


\subsection{Sim_Shen_2010}{

\url{https://doi.org/10.1109/BIBM.2010.5706623}
}


\subsection{Sim_SSDD_2013}{

\url{https://doi.org/10.1016/j.ygeno.2013.04.010}
}


\subsection{Sim_Jiang_1997}{

\url{https://aclanthology.org/O97-1002/}
}


\subsection{Sim_Kappa}{
}


\subsection{Sim_Jaccard}{
}


\subsection{Sim_Dice}{
}


\subsection{Sim_Overlap}{
}
}

\examples{
parents  = c("a", "a", "b", "b", "c", "d")
children = c("b", "c", "c", "d", "e", "f")
annotation = list(
    "a" = 1:3,
    "b" = 3:4,
    "c" = 5,
    "d" = 7,
    "e" = 4:7,
    "f" = 8
)
dag = create_ontology_DAG(parents, children, annotation = annotation)
term_sim(dag, dag_all_terms(dag), "Sim_Ling_1998")
}
