% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_global.R
\docType{data}
\name{ALL_IC_METHODS}
\alias{ALL_IC_METHODS}
\title{All supported IC methods}
\format{
An object of class \code{character} of length 9.
}
\usage{
ALL_IC_METHODS
}
\description{
All supported IC methods
}
\keyword{datasets}
