% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_global.R
\docType{data}
\name{ALL_IC_METHODS}
\alias{ALL_IC_METHODS}
\alias{ALL_TERM_SIM_METHODS}
\alias{ALL_GROUP_SIM_METHODS}
\title{Supported methods}
\format{
An object of class \code{character} of length 9.

An object of class \code{character} of length 34.

An object of class \code{character} of length 28.
}
\usage{
ALL_IC_METHODS

ALL_TERM_SIM_METHODS

ALL_GROUP_SIM_METHODS
}
\value{
A vector all supported methods.
}
\description{
Supported methods
}
\details{
\itemize{
\item \code{ALL_IC_METHODS}: A vector all supported IC methods.
\item \code{ALL_TERM_SIM_METHODS}: A vector all supported term similarity methods.
\item \code{ALL_GROUP_SIM_METHODS}: A vector all supported group similarity methods.
}
}
\examples{
ALL_IC_METHODS
ALL_TERM_SIM_METHODS
ALL_GROUP_SIM_METHODS
}
\keyword{datasets}
