% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_global.R
\name{all_ic_methods}
\alias{all_ic_methods}
\alias{all_term_sim_methods}
\alias{all_group_sim_methods}
\title{Supported methods}
\usage{
all_ic_methods()

all_term_sim_methods()

all_group_sim_methods()
}
\value{
A vector of all supported methods.
}
\description{
Supported methods
}
\details{
\itemize{
\item \code{all_ic_methods()}: A vector of all supported IC methods.
\item \code{all_term_sim_methods()}: A vector of all supported term similarity methods.
\item \code{all_group_sim_methods()}: A vector of all supported group similarity methods.
}
}
\examples{
all_ic_methods()
all_term_sim_methods()
all_group_sim_methods()
}
