% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{term_annotations}
\alias{term_annotations}
\alias{annotated_terms}
\title{Get term-item associations}
\usage{
term_annotations(dag, terms, return = "list")

annotated_terms(dag, anno, return = "list")
}
\arguments{
\item{dag}{An \code{ontology_DAG} object.}

\item{terms}{A vector of term names.}

\item{return}{Whether to return a list or a matrix?}

\item{anno}{A vector of annotated item names.}
}
\value{
A list or a binary matrix showing annotation relations between terms and items.
}
\description{
Get term-item associations
}
\examples{
parents  = c("a", "a", "b", "b", "c", "d")
children = c("b", "c", "c", "d", "e", "f")
annotation = list(
    "a" = 1:3,
    "b" = 3:4,
    "c" = 5,
    "d" = 7,
    "e" = 4:7,
    "f" = 8
)
dag = create_ontology_DAG(parents, children, annotation = annotation)
term_annotations(dag, letters[1:6])
annotated_terms(dag, c("1", "2", "3"))
}
