% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_ancestor.R
\name{MICA_term}
\alias{MICA_term}
\alias{MICA_IC}
\alias{LCA_term}
\alias{LCA_depth}
\alias{NCA_term}
\title{Various types of common ancestors}
\usage{
MICA_term(dag, terms, IC_method, in_labels = TRUE)

MICA_IC(dag, terms, IC_method)

LCA_term(dag, terms, in_labels = TRUE)

LCA_depth(dag, terms)

NCA_term(dag, terms, in_labels = TRUE)
}
\arguments{
\item{dag}{An \code{ontology_DAG} object.}

\item{terms}{A vector of term names.}

\item{IC_method}{An IC method. Valid values are in \link{ALL_IC_METHODS}.}

\item{in_labels}{Whether the terms are in their names or as the integer indices?}
}
\value{
\itemize{
\item \code{MICA_term()} returns a character matrix of the MICA terms.
\item \code{MICA_IC()} returns a numeric matrix of the IC of the MICA terms.
\item \code{LCA_term()} returns a character matrix of the LCA term.
\item \code{LCA_depth()} reutrns an integer matrix of the depth of the LCA terms.
\item \code{NCA_term()} returns a character matrix of the NCA term. The shortest distance can be calculated by \code{\link[=shortest_distances_via_CA]{shortest_distances_via_CA()}}.
}
}
\description{
Various types of common ancestors
}
\details{
There are the following three types of common ancestors:
\itemize{
\item MICA (most informative common ancestor): The common ancestor with the highest IC value.
\item LCA (lowest common ancestor): The common ancestor with the largest height. If there are multiple ancestors having
the max height, the ancestor with the smallest distance to the two terms is taken.
\item NCA (nearest common ancestor): The common ancestor with the smallest distance to the two terms. If there are multiple
ancestors with the same smallest distance, the ancestor with the largest height is taken.
}
}
\examples{
parents  = c("a", "a", "b", "b", "c", "d")
children = c("b", "c", "c", "d", "e", "f")
dag = create_ontology_DAG(parents, children)
MICA_term(dag, letters[1:6], "IC_universal")
MICA_IC(dag, letters[1:6], "IC_universal")
LCA_term(dag, letters[1:6])
LCA_depth(dag, letters[1:6])
NCA_term(dag, letters[1:6])
}
