% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructor.R
\name{create_ontology_DAG}
\alias{create_ontology_DAG}
\title{Create the ontology_DAG object}
\usage{
create_ontology_DAG(
  parents,
  children,
  relations = NULL,
  source = "Ontology",
  annotation = NULL
)
}
\arguments{
\item{parents}{A vector of parent terms.}

\item{children}{A vector of child terms.}

\item{relations}{A vector of parent-child relations.}

\item{source}{Source of the ontology. Simply used as a mark of the object.}

\item{annotation}{A list of items that are annotated to the terms. Names in the list should be the term names. In the DAG, items
annotated to a term will also be annotated to its parents. Such upstreaming merging
is applied automatically in the package.}
}
\value{
An `ontology_DAG` object.
}
\description{
Create the ontology_DAG object
}
