% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructor.R
\name{create_ontology_DAG}
\alias{create_ontology_DAG}
\title{Create the ontology_DAG object}
\usage{
create_ontology_DAG(
  parents,
  children,
  relations = NULL,
  source = "Ontology",
  annotation = NULL
)
}
\arguments{
\item{parents}{A character vector of parent terms.}

\item{children}{A character vector of child terms.}

\item{relations}{A character vector of parent-child relations, e.g. "isa", "part of", or self-defined semantic relations.}

\item{source}{Source of the ontology. It is only used as a mark of the object.}

\item{annotation}{A list of character vectors which contain items annotated to the terms. Names of the list should be the term names. In the DAG, items
annotated to a term will also be annotated to its parents. Such merging
is applied automatically in the package.}
}
\value{
An \code{ontology_DAG} object.
}
\description{
Create the ontology_DAG object
}
\examples{
parents  = c("a", "a", "b", "b", "c", "d")
children = c("b", "c", "c", "d", "e", "f")
dag = create_ontology_DAG(parents, children)

# with annotations
annotation = list(
    "a" = c("t1", "t2", "t3"),
    "b" = c("t3", "t4"),
    "c" = "t5",
    "d" = "t7",
    "e" = c("t4", "t5", "t6", "t7"),
    "f" = "t8"
)
dag = create_ontology_DAG(parents, children, annotation = annotation)

# with relations
dag = create_ontology_DAG(parents, children, 
    relations = c("isa", "part of", "isa", "part of", "isa", "part of"))
}
