% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructor.R
\name{create_ontology_DAG_from_GO_db}
\alias{create_ontology_DAG_from_GO_db}
\title{Create the ontology_DAG object from the GO.db package}
\usage{
create_ontology_DAG_from_GO_db(
  namespace = "BP",
  relations = c("isa", "part of"),
  org_db = NULL
)
}
\arguments{
\item{namespace}{One of "BP", "CC" and "MF".}

\item{relations}{Type of the GO term relations.}

\item{org_db}{The name of the organism package or the corresponding database object, e.g. `"org.Hs.eg.db"` or 
directly the [org.Hs.eg.db::org.Hs.eg.db] object for human. The gene annotation to GO terms will be added
to the object.}
}
\value{
An `ontology_DAG` object.
}
\description{
Create the ontology_DAG object from the GO.db package
}
\examples{
\dontrun{
dag = create_ontology_DAG_from_GO_db()
dag = create_ontology_DAG_from_GO_db("CC", 
    relations = "isa", org_db = "org.Hs.eg.db")
}
}
