% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructor.R
\name{create_ontology_DAG_from_GO_db}
\alias{create_ontology_DAG_from_GO_db}
\title{Create the ontology_DAG object from the GO.db package}
\usage{
create_ontology_DAG_from_GO_db(
  namespace = "BP",
  relations = "part of",
  org_db = NULL,
  evidence_code = NULL,
  retrieve_alternative = FALSE,
  verbose = simona_opt$verbose
)
}
\arguments{
\item{namespace}{One of "BP", "CC" and "MF".}

\item{relations}{Types of the GO term relations. In the \strong{GO.db} package, the GO term relations can be "is_a", "part_of",
"regulates", "negatively regulates", "positively regulates". Note since "regulates" is a parent relation
of "negatively regulates", "positively regulates", if "regulates" is selected, "negatively regulates" and "positively regulates"
are also selected. Note "is_a" is always included.}

\item{org_db}{The name of the organism package or the corresponding database object, e.g. \code{"org.Hs.eg.db"} or
directly the \code{\link[org.Hs.eg.db:org.Hs.egBASE]{org.Hs.eg.db::org.Hs.eg.db}} object for human, then the gene annotation to GO terms will be added
to the object. For other non-model organisms, consider to use the \strong{AnnotationHub} package to find one.}

\item{evidence_code}{A vector of evidence codes for gene annotation to GO terms. See \url{https://geneontology.org/docs/guide-go-evidence-codes/}.}

\item{retrieve_alternative}{Whether to retrieve alternative/obsolete GO terms from geneontology.org?}

\item{verbose}{Whether to print messages.}
}
\value{
An \code{ontology_DAG} object.
}
\description{
Create the ontology_DAG object from the GO.db package
}
\examples{
dag = create_ontology_DAG_from_GO_db()
dag
}
