% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transverse.R
\name{dag_depth}
\alias{dag_depth}
\alias{dag_height}
\alias{dag_dist_from_root}
\alias{dag_dist_from_leaves}
\title{Depth and height of the DAG}
\usage{
dag_depth(dag, use_cache = TRUE)

dag_height(dag, use_cache = TRUE)

dag_dist_from_root(dag, use_cache = TRUE)

dag_dist_from_leaves(dag, use_cache = TRUE)
}
\arguments{
\item{dag}{An \code{ontology_DAG} object.}

\item{use_cache}{Internally used.}
}
\value{
An integer vector.
}
\description{
Depth and height of the DAG
}
\details{
The depth of a term in the DAG is defined as the maximal distance from the root. The height
of a term in the DAG is the maximal finite distance to the leaf nodes.

\code{dag_dist_from_root()} and \code{dag_dist_from_leaves()} calculate the minimal distance to the root or to the leaves.
}
\examples{
parents  = c("a", "a", "b", "b", "c", "d")
children = c("b", "c", "c", "d", "e", "f")
dag = create_ontology_DAG(parents, children)
dag_depth(dag)
dag_height(dag)
dag_dist_from_root(dag)
dag_dist_from_leaves(dag)
}
