% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructor.R
\name{dag_filter}
\alias{dag_filter}
\title{Filter the DAG}
\usage{
dag_filter(dag, terms = NULL, relations = NULL, root = NULL, leaves = NULL)
}
\arguments{
\item{dag}{A `OntologyDAG` object.}

\item{terms}{A vectof of term names.}

\item{relations}{A vector of relations.}

\item{root}{A vector of root terms. Only their offspring terms are kept.}

\item{leaves}{A vector of leaf terms. Only their ancestor terms are kept.}
}
\description{
Filter the DAG
}
