% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{dag_circular_viz}
\alias{dag_circular_viz}
\alias{dag_graphviz}
\title{Visualize the DAG}
\usage{
dag_circular_viz(dag)

dag_graphviz(dag)
}
\arguments{
\item{dag}{An \code{ontology_Dag} object.}
}
\description{
Visualize the DAG
}
\details{
\code{dag_circular_viz()} uses a circular layout for visualizing large DAGs. \code{dag_graphviz()}
uses a hierarchical layout for visualizing small DAGs.
}
