% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{group_sim}
\alias{group_sim}
\title{Semantic similarity between two groups of terms}
\usage{
group_sim(dag, group1, group2, method, sim_method, control = list())
}
\arguments{
\item{dag}{An \code{ontology_DAG} object.}

\item{group1}{A vector of term names.}

\item{group2}{A vector of term names.}

\item{method}{A group similarity method. All available methods are in \code{\link{ALL_GROUP_SIM_METHODS}}.}

\item{sim_method}{A Term similarity method. All available methods are in \code{\link{ALL_TERM_SIM_METHODS}}.}

\item{control}{A list of parameters passing to individual methods. See the subsections.}
}
\value{
A numeric scalar.
}
\description{
Semantic similarity between two groups of terms
}
\details{
If \code{annotation} is set in \code{create_ontology_DAG()} and you want to directly calculate semantic similarity between two
annotated items, you can first get the associated terms of the two items by \code{\link[=annotated_terms]{annotated_terms()}}:

\if{html}{\out{<div class="sourceCode">}}\preformatted{group1 = annotated_terms(dag, item1)
group2 = annotated_terms(dag, item2)
group_sim(dag, group1, group2, ...)
}\if{html}{\out{</div>}}
}
\section{method}{

what is GroupSim_pairwise_avg


what is GroupSim_pairwise_max


what is GroupSim_pairwise_BMM


what is GroupSim_pairwise_ABM


what is GroupSim_pairwise_HDF


what is GroupSim_pairwise_VHDF


what is GroupSim_pairwise_Froehlich_2007


what is GroupSim_pairwise_Joeng_2014


what is GroupSim_SimALN


what is GroupSim_SimINT


what is GroupSim_spgk


what is GroupSim_SimGIC


what is GroupSim_SimDIC


what is GroupSim_SimUIC


what is GroupSim_SimUI


what is GroupSim_SimDB


what is GroupSim_SimUB


what is GroupSim_SimNTO


what is GroupSim_SimCOU


what is GroupSim_SimCOT


what is GroupSim_SimLP


what is GroupSim_Ye_2005


what is GroupSim_Cho_2007


what is GroupSim_SimALD


what is GroupSim_Jaccard


what is GroupSim_Dice


what is GroupSim_Overlap


what is GroupSim_Kappa
}

\examples{
parents  = c("a", "a", "b", "b", "c", "d")
children = c("b", "c", "c", "d", "e", "f")
annotation = list(
    "a" = 1:3,
    "b" = 3:4,
    "c" = 5,
    "d" = 7,
    "e" = 4:7,
    "f" = 8
)
dag = create_ontology_DAG(parents, children, annotation = annotation)
group_sim(dag, c("c", "e"), c("d", "f"), 
    method = "GroupSim_pairwise_avg", 
    sim_method = "Sim_Ling_1998"
)
}
