% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_obo}
\alias{import_obo}
\title{Import .obo file to an ontology_DAG object}
\usage{
import_obo(file)
}
\arguments{
\item{file}{Path of the .obo file or an URL. Pass to \code{\link[base:readLines]{base::readLines()}}.}
}
\value{
An \code{ontology_DAG} object.
}
\description{
Import .obo file to an ontology_DAG object
}
\details{
Public bio-ontologies can be obtained from \href{http://obofoundry.org/}{Ontology Foundry} or \href{https://bioportal.bioontology.org/}{BioPortal} where
.obo files for the corresponding ontologies can be easily found. If there is no such .obo format file, the .owl file
can be converted to the .obo file using \href{http://robot.obolibrary.org/}{the \strong{ROBOT} tool}. An example command is:

\if{html}{\out{<div class="sourceCode">}}\preformatted{robot convert --input ontology.owl --output -o ontology.obo --check false
}\if{html}{\out{</div>}}
}
\examples{
\dontrun{
# The plant ontology: http://obofoundry.org/ontology/po.html 
dag = import_obo("https://raw.githubusercontent.com/Planteome/plant-ontology/master/po.obo")
}
}
