% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_ancestor.R
\name{MICA_term}
\alias{MICA_term}
\alias{MICA_IC}
\alias{LCA_term}
\alias{LCA_depth}
\alias{NCA_term}
\title{Most informative common ancestor (MICA) and lowest common ancestor (LCA)}
\usage{
MICA_term(dag, terms, IC_method, in_labels = TRUE)

MICA_IC(dag, terms, IC_method)

LCA_term(dag, terms, in_labels = TRUE)

LCA_depth(dag, terms)

NCA_term(dag, terms, in_labels = TRUE)
}
\arguments{
\item{dag}{A `ontology_DAG` object.}

\item{terms}{A vector of term names.}

\item{IC_method}{An IC method. Valid values are in [ALL_IC_METHODS].}

\item{in_labels}{Whether the terms are in their names or as the integer indices?}
}
\value{
- `MICA_term()` returns a character matrix of the MICA terms. 
- `MICA_IC()` returns a numeric matrix of the IC of the MICA terms.
- `LCA_term()` returns a character matrix of the LCA term.
- `LCA_depth()` reutrns an integer matrix of the depth of the LCA terms.
- `NCA_term()` returns a character matrix of the nearest common ancestors.
}
\description{
Most informative common ancestor (MICA) and lowest common ancestor (LCA)
}
