% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{n_annotations}
\alias{n_annotations}
\title{Number of annotated items}
\usage{
n_annotations(dag, use_cache = TRUE)
}
\arguments{
\item{dag}{An \code{ontology_DAG} object.}

\item{use_cache}{Internally used.}
}
\value{
An integer vector.
}
\description{
Number of annotated items
}
\details{
Due to the nature of DAG, a parent term includes all annotated items of its child terms. This upstream-merging
is automatically applied in the function.
}
\examples{
parents  = c("a", "a", "b", "b", "c", "d")
children = c("b", "c", "c", "d", "e", "f")
annotation = list(
    "a" = 1:3,
    "b" = 3:4,
    "c" = 5,
    "d" = 7,
    "e" = 4:7,
    "f" = 8
)
dag = create_ontology_DAG(parents, children, annotation = annotation)
n_annotations(dag)
}
