% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{n_annotations}
\alias{n_annotations}
\title{Number of annotated items}
\usage{
n_annotations(dag, use_cache = TRUE)
}
\arguments{
\item{dag}{A `ontology_DAG` object.}

\item{use_cache}{Internally used.}
}
\value{
An integer vector.
}
\description{
Number of annotated items
}
\details{
`annotation` argument should be set in [create_ontology_DAG()].
}
\examples{
\dontrun{
dag = create_ontology_DAG_from_GO_db(org_db = "org.Hs.eg.db")
head(n_annotations(dag))
}
}
