% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructor.R
\name{[,ontology_DAG-method}
\alias{[,ontology_DAG-method}
\alias{[[,ontology_DAG,character,missing-method}
\title{Create sub-DAGs}
\usage{
\S4method{[}{ontology_DAG}(x, i, j, ..., drop = FALSE)

\S4method{[[}{ontology_DAG,character,missing}(x, i, j, ...)
}
\arguments{
\item{x}{An \code{ontology_DAG} object.}

\item{i}{A single term name. The value should be a character vector. This corresponds to the roots.}

\item{j}{A single term name. The value should be a character vector. This corresponds to the leaves.}

\item{...}{Ignored.}

\item{drop}{Ignored.}
}
\value{
An \code{ontology_DAG} object.
}
\description{
Create sub-DAGs
}
\details{
It returns a sub-DAG taking node \code{i} as the root and \code{j} as the leaves.
}
\examples{
parents  = c("a", "a", "b", "b", "c", "d")
children = c("b", "c", "c", "d", "e", "f")
dag = create_ontology_DAG(parents, children)
dag["b"]
dag[["b"]]
dag["b", "f"]
dag[, "f"]
}
