% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{term_IC}
\alias{term_IC}
\title{Information content}
\usage{
term_IC(dag, method, control = list())
}
\arguments{
\item{dag}{An `ontology_DAG` object.}

\item{method}{An IC method. All available methods are in [ALL_IC_METHODS].}

\item{control}{A list of parameters passing to individual methods.}
}
\value{
A numeric vector.
}
\description{
Information content
}
\details{
When using `IC_annotation`, `annotation` should be already set in [create_ontology_DAG()].
}
\examples{
parents  = c("a", "a", "b", "b", "c", "d")
children = c("b", "c", "c", "d", "e", "f")
dag = create_ontology_DAG(parents, children)
for(method in ALL_IC_METHODS) {
    cat("===", method, "====\n")
    print(term_IC(dag, method))
    cat("\n")
}
}
