% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{term_IC}
\alias{term_IC}
\title{Information content}
\usage{
term_IC(dag, method, control = list())
}
\arguments{
\item{dag}{An \code{ontology_DAG} object.}

\item{method}{An IC method. All available methods are in \link{ALL_IC_METHODS}.}

\item{control}{A list of parameters passing to individual methods.}
}
\value{
A numeric vector.
}
\description{
Information content
}
\section{method}{

what is IC_annotation


what is IC_universal


what is IC_Zhang_2006


what is IC_Seco_2004


what is IC_Zhou_2008


what is IC_Seddiqui_2010


what is IC_Sanchez_2011


what is IC_Meng_2012


what is IC_Wang_2007
}

\examples{
parents  = c("a", "a", "b", "b", "c", "d")
children = c("b", "c", "c", "d", "e", "f")
annotation = list(
    "a" = 1:3,
    "b" = 3:4,
    "c" = 5,
    "d" = 7,
    "e" = 4:7,
    "f" = 8
)
dag = create_ontology_DAG(parents, children, annotation = annotation)
term_IC(dag, "IC_annotation")
}
