% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{term_sim}
\alias{term_sim}
\title{Semantic similarity}
\usage{
term_sim(dag, terms, method, control = list())
}
\arguments{
\item{dag}{An `ontology_DAG` object.}

\item{terms}{A vector of term names.}

\item{method}{A similarity method. All available methods are in [ALL_TERM_SIM_METHODS].}

\item{control}{A list of parameters passing to individual methods.}
}
\value{
A numeric matrix.
}
\description{
Semantic similarity
}
\examples{
parents  = c("a", "a", "b", "b", "c", "d")
children = c("b", "c", "c", "d", "e", "f")
annotation = list(
    "a" = 1:3,
    "b" = 3:4,
    "c" = 5,
    "d" = 7,
    "e" = 4:7,
    "f" = 8
)
dag = create_ontology_DAG(parents, children, annotation = annotation)
term_sim(dag, dag_all_terms(dag), "Sim_Ling_1998")
}
